==> analyses/3_kid_vocabs/1_semantic_space_by_kid_analyses.Rmd <==
---
title: Semantic space by-kid analyses
date: "`r Sys.Date()`"
output:
  html_document:
    highlight: tango
    toc: true
    number_sections: false
    theme: cerulean
    toc_float: false
    code_folding: hide
---

```{r setup, message = F, warning = F}
library(knitr)

opts_chunk$set(echo = T, message = F, warning = F, 
               error = F, cache = F, tidy = F)

library(tidyverse)
library(langcog)
library(data.table)
library(feather)

theme_set(theme_classic(base_size = 10))
```

*QUESTION*: Do the distance properties of a kids vocab at t1 predict vocabulary outcomes at t2?

For each kids vocab, getting coordinates of each word in semantic space using fasttext english wikipedia model. The predictors are the mean, median, var distance between words for each kid.

Vocabulary outcome measures: mtld at t2, mtld change, number of trigrams at t2 and frequency of trigrams at t1

In previous versions of this analysis I transformed the 300-D wikipedia model into tsne coordinates for the words in all kids' vocabs. This analysis suggested an effect whereby vocabs with larger mean distance at t1 predicted mtld at t2. However, the tsne coordinates were probabilistic (changed with each run of the model), and thus these results were not consistent.

So, here I'm doing distance calculations using cosine on the full 300-D space. The effects depend on the cuttoff for counting a kid as knowing a word, with larger effects for bigger cuttoffs. 

Broadly, the pattern of findings is that vocabularies with high distance and low variability have high mtld at t2. This is also true for number of trigrams. For trigram frequency, the opposite is true: low distance high variability predicts more frequent trigrams at t2.

```{r, params}
MINWORDSFORVOCAB <- 1
```

The min words for vocab here is `r MINWORDSFORVOCAB`.

## Get df with all measures
Read in data
```{r}
all_types <- read_csv("../1_mtld_measure/data/target_types_for_MTLD_kids_600_900.csv") 
fasttext_model <- read_feather("fast_text_childes_words_600_900.feather")
groups_info <- read_csv("../1_mtld_measure/data/groups_info_600_900_corrected.csv")
trigrams <- read_csv("../2_trigrams/mtld_continuous_trigram_by_kid_MIN1.csv")
```

Get filtered version of types for each kid
```{r}
types_clean <- all_types %>%
  filter(tbin == "t1") %>%
  mutate(gloss_clean = tolower(gloss))   %>%
  group_by(target_child_id, gloss_clean) %>%
  summarize(count = sum(count)) %>%
  filter(count >= MINWORDSFORVOCAB)
```

Get vocab measures by kids
```{r}
get_vocab_measure_by_kid <- function(id, data, model){
  this_kids_model <- model %>%
    filter(target_word %in% data$gloss_clean)
 
   # get pairwise distances
   all_dists = coop::cosine(t(this_kids_model[,-1]))
  
   data.frame(target_child_id = id,
              mean_dist_t1 = mean(all_dists),
              median_dist_t1 = median(all_dists),
              var_dist_t1 =  ifelse(mean(var(all_dists)) == 0, NA, mean(var(all_dists))),
              n_t1 = nrow(all_dists)) 
}

nested_data_by_kid <- nest(types_clean, -target_child_id)

vocab_measures <- map2_df(nested_data_by_kid$target_child_id, 
                          nested_data_by_kid$data, 
                          get_vocab_measure_by_kid, 
                          fasttext_model)
```

Merge in other variables
```{r}
other_measures <- read_csv("semantic_density_df.csv")

vocab_df <- vocab_measures %>%
  left_join(groups_info %>% select(delta_resid_group, target_child_id, mtld_t1, 
                                          mtld_t2, age_t1, age_t2, mtld_diff, age_diff)) %>%
  mutate(log_mtld_t2 = log(mtld_t2 + 1),
         log_mtld_t1 = log(mtld_t1 + 1),
         log_median_dist_t1 = log(median_dist_t1),
         log_mean_dist_t1 = log(mean_dist_t1),
         log_var_dist_t1 = log(var_dist_t1),
         log_n_t1 = log(n_t1)) %>%
  left_join(trigrams %>% select(target_child_id, 
                                log_num_trigrams_t1,
                                log_num_trigrams_t2,
                                mean_log_freq_trigrams_t1,
                                mean_log_freq_trigrams_t2)) %>%
  mutate_if(is.numeric, scale) # scale everything for regressions 
```

## Regressions
### Predicting MTLD at t2

```{r}
lm(log_mtld_t2 ~ log_median_dist_t1 + age_t1 + age_diff + log_mtld_t1 + log_n_t1,
   data = vocab_df) %>%
  summary()

lm(log_mtld_t2 ~ log_mean_dist_t1 + age_t1 + age_diff + log_mtld_t1 + log_n_t1,
     data = vocab_df) %>%
    summary()

lm(log_mtld_t2 ~ log_var_dist_t1 + age_t1 + age_diff + log_mtld_t1 + log_n_t1,
   data = vocab_df) %>%
  summary()

lm(log_mtld_t2 ~ log_median_dist_t1+ log_var_dist_t1 + age_t1 + age_diff + log_mtld_t1 + log_n_t1,
   data = vocab_df) %>%
  summary()
```

### Predicting number of trigrams at t2
```{r}
lm(log_num_trigrams_t2 ~ log_median_dist_t1 + age_t1 + age_t2  + log_n_t1 + log_num_trigrams_t1,
   data = vocab_df) %>%
  summary()

lm(log_num_trigrams_t2 ~ log_mean_dist_t1 + age_t1 + age_t2  + log_n_t1 + log_num_trigrams_t1,
   data = vocab_df) %>%
  summary()


lm(log_num_trigrams_t2 ~ log_var_dist_t1  + age_t1 + age_t2  + log_num_trigrams_t1 + log_n_t1 ,
   data = vocab_df) %>%
  summary()

lm(log_num_trigrams_t2 ~ log_mean_dist_t1  +  log_var_dist_t1  + age_t1 + age_t2  + log_num_trigrams_t1 + log_n_t1 ,
   data = vocab_df) %>%
  summary()

```

### Predicting frequency of trigrams at t2

```{r}
lm(mean_log_freq_trigrams_t2 ~ log_median_dist_t1 + age_t1 + age_t2  + log_n_t1 + mean_log_freq_trigrams_t1,
   data = vocab_df) %>%
  summary()

lm(mean_log_freq_trigrams_t2 ~ log_mean_dist_t1 + age_t1 + age_t2  + log_n_t1 + mean_log_freq_trigrams_t1,
   data = vocab_df) %>%
  summary()

lm(mean_log_freq_trigrams_t2 ~ log_var_dist_t1  + age_t1 + age_t2  + mean_log_freq_trigrams_t1 + log_n_t1 ,
   data = vocab_df) %>%
  summary()

lm(mean_log_freq_trigrams_t2 ~ log_median_dist_t1  +  log_var_dist_t1  + age_t1 + age_t2  + mean_log_freq_trigrams_t1 + log_n_t1 ,
   data = vocab_df) %>%
  summary()

```
